The admin search view is no longer available. Use phoenix instead.
The admin search view should not have been available to search orders.
This was abused in the storefronts to create the my orders pages.

There was a bug in isaac that let this happen and now it has been fixed.
So this search view is no longer available to customer JWTs.

Unfortunately this will return a different payload than the search view.
This requires fixing the storefronts.